requirements.txt
streamlit
pandas
scikit-learn
numpy
matplotlib
seaborn
scipy
scikit-surprise